+++ tool/build_release.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# TODO(kenz): delete this script once we can confirm it is not used in the
+# Dart SDK or in infra tooling.
+
 # Copyright 2019 The Chromium Authors. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
@@ -38,7 +41,7 @@ echo "Flutter Path: $(which flutter)"
 echo "Flutter Version: $(flutter --version)"
 
 if [[ $1 = "--update-perfetto" ]]; then
-  $TOOL_DIR/update_perfetto.sh
+  devtools_tool update-perfetto
 fi
 
 pushd $DEVTOOLS_DIR/packages/devtools_shared